bench(dna_bench): allow fractional --tps target
Previously --tps was parsed via atoi(), so any fractional input was
silently truncated (--tps 1.5 → 1). That blocks bisecting where the
cluster cliff lies: real-cluster runs 2026-05-03 showed clean
sustained behaviour at tps=1 (0.75 effective, 90% commit, 4.97 s
median latency) and a hard collapse at tps=2 (0.013 effective, 1%
commit, 100-wallet burst overflowing the 64-slot mempool).

To resolve the gap between 1 and 2 the round_budget formula needs
intermediate values:

  round_budget_us = 1e6 * pool / tps_target

Switch tps_target from int to double in struct dna_bench_run_cfg
and the per-thread args mirror in dna_bench_native.c. Parse with
strtod, format JSON / stderr with %g (renders 1 / 1.5 / 2 cleanly,
no trailing zeros).

The fork/exec sustained loop in dna_bench_run.c keeps its int
target_tps internally — narrowing happens at the call into
run_loop_until_deadline. Sustained mode in production goes through
the in-process libnodus path (dna_bench_native.c) which now
honours the full double precision in the round-budget formula.

+++ nodus/tests/bench/dna_bench_internal.h
@@ -108,7 +108,7 @@ enum dna_bench_mode {
 
 struct dna_bench_run_cfg {
     enum dna_bench_mode mode;
-    int      tps_target;
+    double   tps_target;     /* fractional rates allowed (e.g. 1.5) */
     int      duration_s;
     int      ramp_start;
     int      ramp_end;
